Move IsDirectoryEmpty out of utils:files
Bug: b/543585591
