+++ base/cvd/cuttlefish/files/is_directory_empty.cc
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2017 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "cuttlefish/files/is_directory_empty.h"
+
+#include <dirent.h>
+#include <errno.h>
+
+#include <memory>
+#include <string>
+
+#include "cuttlefish/posix/strerror.h"
+#include "cuttlefish/result/expect.h"
+#include "cuttlefish/result/result_type.h"
+
+namespace cuttlefish {
+
+Result<bool> IsDirectoryEmpty(const std::string& path) {
+  std::unique_ptr<DIR, int (*)(DIR*)> direc(opendir(path.c_str()), closedir);
+  CF_EXPECTF(direc.get(), "opendir('{}') failed: {}", path, StrError(errno));
+
+  int cnt = 0;
+  while (::readdir(direc.get())) {
+    cnt++;
+    if (cnt > 2) {
+      return false;
+    }
+  }
+  return true;
+}
+
+}  // namespace cuttlefish
