+++ base/cvd/cuttlefish/host/commands/assemble_cvd/BUILD.bazel
@@ -34,6 +34,7 @@ cf_cc_binary(
         "//cuttlefish/files:directory_contents",
         "//cuttlefish/files:directory_exists",
         "//cuttlefish/files:file_exists",
+        "//cuttlefish/files:is_directory_empty",
         "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/flag_parser",
         "//cuttlefish/host/commands/assemble_cvd:assemble_cvd_flags",
@@ -325,6 +326,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:network",
         "//cuttlefish/files:directory_exists",
         "//cuttlefish/files:file_exists",
+        "//cuttlefish/files:is_directory_empty",
         "//cuttlefish/host/commands/assemble_cvd:alloc",
         "//cuttlefish/host/commands/assemble_cvd:assemble_cvd_flags",
         "//cuttlefish/host/commands/assemble_cvd:disk_image_flags_vectorization",
